sus_connection/requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.4.0
crispy-bootstrap3==2024.1
cv-3==1.2.0
defusedxml==0.7.1
Django==5.1.3
django-bootstrap-icons==0.9.0
django-braces==1.16.0
django-crispy-forms==2.3
django-icons==24.4
django-icons-bootstrap-icons==1.0.1
idna==3.10
numpy==2.1.3
opencv-python==4.10.0.84
pillow==11.0.0
requests==2.32.3
sqlparse==0.5.2
tzdata==2024.2
urllib3==2.2.3
